+++ src/mpflash/mpflash/ask_input.py
@@ -12,7 +12,7 @@ from typing import Dict, List, Sequence, Tuple, Union
 from loguru import logger as log
 
 from mpflash.config import config
-from mpflash.mpboard_id.api import known_stored_boards, local_mp_ports
+from mpflash.mpboard_id import known_stored_boards, local_mp_ports
 from mpflash.mpremoteboard import MPRemoteBoard
 from mpflash.vendor.versions import micropython_versions
