+++ src/mpflash/mpflash/cli_flash.py
@@ -4,7 +4,7 @@ import rich_click as click
 from loguru import logger as log
 
 from mpflash.errors import MPFlashError
-from mpflash.mpboard_id.api import find_stored_board
+from mpflash.mpboard_id import find_stored_board
 from mpflash.vendor.versions import clean_version
 
 from .ask_input import FlashParams, ask_missing_params
