+++ src/mpflash/mpflash/worklist.py
@@ -9,7 +9,7 @@ from mpflash.errors import MPFlashError
 from .config import config
 from .downloaded import find_downloaded_firmware
 from .list import show_mcus
-from .mpboard_id.api import find_stored_board
+from .mpboard_id import find_stored_board
 from .mpremoteboard import MPRemoteBoard
 
 # #########################################################################################################
